The fact that TON_ARCH was set to -march=native by default was causing issue while building for non-standard architectures. TON_ARCH basicly adds it's value to the -march in CFLAGS/CXXFLAGS. native forces compiler to atempt to CPU specific optimizations. if -march is not set, then general arch and CPU features is populated by cmake.
Problem is that native is not always supported but current ton build
scenarios do not take it into account.

In context of this container i'll just set it to cmake defined since
it's for dev purposes and optimizations are irrelevant.

However this has to be addressed by the ton cmake buildsystem scenarios.